@@ -34,22 +34,23 @@
   "license": "MIT",
   "browserslist": "> 0.25%, not dead",
   "engines": {
-    "node": ">=10.0.0",
+    "node": ">=14.0.0",
     "npm": ">=6.0.0"
   },
   "dependencies": {
+    "@emotion/react": "^11.5.0",
+    "@emotion/styled": "^11.3.0",
+    "@interslavic/razumlivost": "^1.2.0",
     "@interslavic/steen-utils": "^1.1.1",
-    "@material-ui/core": "^4.11.3",
-    "@material-ui/icons": "^4.11.2",
-    "@material-ui/lab": "^4.0.0-alpha.57",
+    "@mui/icons-material": "^5.0.4",
+    "@mui/material": "^5.0.4",
     "core-js": "^3.10.2",
     "gas-client": "^0.1.0",
     "lodash": "^4.17.21",
     "prop-types": "^15.7.2",
-    "react": "^16.13.1",
-    "react-bootstrap": "^1.3.0",
-    "react-dom": "^16.13.1",
-    "react-transition-group": "^4.4.1"
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "react-transition-group": "^4.4.2"
   },
   "devDependencies": {
     "@babel/cli": "^7.10.5",
