@@ -1,5 +1,5 @@
 import React from 'react';
-import { Box, CircularProgress } from '@material-ui/core';
+import { Box, CircularProgress } from '@mui/material';
 
 export default function CircularProgressBox() {
   return (
