@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { Card } from '@material-ui/core';
+import { Card } from '@mui/material';
 import Pre from './Pre';
 
 type ErrorBoundaryState = {
