@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import Sidebar from './containers/Sidebar';
-import ErrorBounary from './components/auxiliary/ErrorBounary';
+import Sidebar from './common/Sidebar';
+import ErrorBounary from './common/components/ErrorBounary';
 
 ReactDOM.render(
   <ErrorBounary>
