@@ -1,21 +1,21 @@
 import { SheetType } from './SheetType';
 
-export default function buildSheetName(
-  sheetType: SheetType,
-  languageCode?: string
-) {
-  switch (sheetType) {
-    case 'vocabulary':
-      return `Vocabulary [ISV]`;
-    case 'flavorization':
-      return `Flavorization [${languageCode.toUpperCase()}]`;
-    case 'translation':
-      return `Translations [${languageCode.toUpperCase()}]`;
-    case 'i18n':
-      return `I18N`;
-    default:
-      throw new Error(
-        `Cannot build sheet name for unsupported sheet type: ${sheetType}`
-      );
+export default function buildSheetName(sheetType: SheetType, arg?: string) {
+  if (sheetType === 'dictionary') {
+    return `words`;
   }
+
+  if (sheetType === 'i18n') {
+    return `i18n`;
+  }
+
+  if (sheetType === 'flavorization' && arg) {
+    return `rules [${arg.toUpperCase()}]`;
+  }
+
+  if (sheetType === 'translation' && arg) {
+    return `translations ${arg.toLowerCase()}]`;
+  }
+
+  throw new Error(`Cannot build sheet name for (${sheetType}, ${arg})`);
 }
